% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finance analysis.R
\name{Spend_analysis}
\alias{Spend_analysis}
\title{Spend Analyses}
\usage{
Spend_analysis(input, month, store)
}
\arguments{
\item{input}{financial data output from "Parse pdf download" function
example July_Amazon <- spend_analysis(demodata2,7,"Amazon")}

\item{month}{month to analyze entered in numeric without preceeding zeros}

\item{store}{store or description keyword entered in quotation marks}
}
\description{
This function filters through finance data to gather spending of different months at different stores
}
