% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize spend.R
\name{Visualize_spend}
\alias{Visualize_spend}
\title{Visualize Spend}
\usage{
Visualize_spend(input_df, X, Y)
}
\arguments{
\item{X}{names of categories analyzed}

\item{Y}{amounts spent}

\item{input}{data frame with at least column X and column Y defined below}
}
\description{
This function produces a plot comparing spending of different stores
}
